sdk: refactoring of the client credentials and property names
